+++ doc/ChangeLog.md
@@ -17,8 +17,9 @@ All notable changes to the project are documented in this file.
    RPC and add boot order to operational datastore.
  - SSH Server is now configurable, issue #441
    SSH Server and NETCONF Server now uses the same SSH hostkey in factory-config
- - Add support for GRE/GRETAP tunnels
+ - Support for GRE/GRETAP tunnels
  - Support for STP/RSTP on bridges
+ - Support for VXLAN tunnels
 
 ### Fixes